* Added the Context class to store usefull informations for the rendering process * Moved rng into Context * Fixed the gamma used in the programm. The Color class now stores the correct values * Replaced CheckeredMaterial with MaskedMaterial which allows to blend any two materials together using a given mask function

+++ Code/SPath/src/kaloffl/spath/math/Vec3d.scala
@@ -76,7 +76,7 @@ case class Vec3d(x: Double, y: Double, z: Double) {
     return Vec3d(nx / length, ny / length, nz / length)
   }
 
-  def randomHemisphere(random: () ⇒ Float): Vec3d = {
+  def randomHemisphere(random: () ⇒ Double): Vec3d = {
     val angle = random() * 2.0 * Math.PI
     val rnd = random()
     val dist = Math.sqrt(1.0 - rnd * rnd)
